Remove RecallCutoff protocol
