+++ docs/ARCHITECTURE.md
@@ -126,7 +126,7 @@ form `ARCHITECTURE.md §N` is used, sparingly.
 | §8.4 Streams & attempt exposure | `LLMEventStream`, `LLMStreamTermination`, `ProviderFailure`, `ProviderFailureAccounting`, `ProviderInferenceCancellation`, `HTTPStreamExchange`, `HTTPTransportFailure`, `BoundedAsyncChannel` (ClawCore); `ProviderAttemptExposure`, `ChatGPTResponsesAttemptEngine` (ClawLLM) |
 | §8.5 Provider replay state | `ProviderExchangeState` (ClawCore); `ChatGPTProviderStateCodec` (ClawLLM) |
 | §8.6 Reliability & cost | `ProviderCallID`, `ProviderCallIDGenerating`, `LLMCostPolicy`, `LLMInputReservationPolicy` (ClawCore); `OpenAICompatibleProvider`, `SSEParser`, `PriceFileLoader` (ClawLLM) |
-| §9 Memory & context | `ContextBuilder`, `BudgetFitter`, `DropMarker`, `MemoryRanker`, `RecallCutoff`, `HistoryHygiene`, `LabeledContextFactory` (ClawAgent); `WorkspaceReading`, `WorkspaceSkills`, `SkillDescriptor`, `SkillScanResult`, `WorkspaceWarning`, `FrontmatterFence` (ClawCore); `FileSystemWorkspace` (ClawWorkspace); `SkillDiagnostics` (ClawGateway) |
+| §9 Memory & context | `ContextBuilder`, `BudgetFitter`, `DropMarker`, `MemoryRanker`, `CandidateCapRecallCutoff`, `HistoryHygiene`, `LabeledContextFactory` (ClawAgent); `WorkspaceReading`, `WorkspaceSkills`, `SkillDescriptor`, `SkillScanResult`, `WorkspaceWarning`, `FrontmatterFence` (ClawCore); `FileSystemWorkspace` (ClawWorkspace); `SkillDiagnostics` (ClawGateway) |
 | §10 Tool system & policy | `ToolRegistry`, `FileReadTool`, `FileWriteTool`, `MemoryWriteTool`, `WebFetchTool`, `WebSearchTool`, `SkillLoadTool` (ClawTools); `Tool`, `ToolDefinition`, `RiskLevel`, `ToolDispatching`, `ToolFenceLabels`, `WorkspacePathContainment` (ClawCore) |
 | §10.3 MCP client | `MCPStreamableHTTPTransport`, `MCPServerSession`, `MCPCatalogResolver`, `ResolvedMCPCatalog`, `MCPServerOutcome`, `MCPMetadataSanitizer`, `MCPTool`, `MCPToolNamer`, `MCPSchemaNormalizer`, `MCPDescriptionCap`, `MCPValueBridge`, `MCPDiscoveryLimits`, `MCPTransportLimits`, `MCPNegotiatingTransport` (ClawMCP); `MCPConfig`, `MCPServerConfig`, `MCPToolFilter`, `MCPHTTPHeader`, `MCPLimits`, `MCPNaming`, `MCPConfigError` (ClawCore); `MCPConfigLoader` (ClawWorkspace); `EncryptedMCPCredentialStore`, `MCPCredentialLoad`, `SealedCredentialFile` (ClawSecrets); `MCPBootInputs`, `MCPProbe`, `MCPSessionFactory`, `MCPDoctorRows`, `MCPCommand`, `MCPSessionLifecycleService`, `DaemonBuilder.resolveMCPStack` (clawd) |
 | §11 Approval system | `Approval`, `ApprovalFSM`, `PendingToolAction`, `RecordedToolAction` (ClawCore); `ApprovalStoreGRDB` (ClawData); the §6.2/§6.5 gateway symbols above |
